FWQCX/requirements.txt
Flask==2.3.3
Flask-Cors==4.0.0
gunicorn==20.1.0
psycopg2-binary==2.9.9
cryptography==41.0.7
requests==2.31.0
python-dotenv==1.0.0